Fix iOS chat CI lint gates
